Enhance error reporting and documentation.
I ran into a number of annoying warts when using ``goflymake`` on
Mac OS X with the Cocoa version of Emacs 24.3, whereby a number of
ancilliary environment setup foo was required.  Tracking down the
anomalies was a bit onerous, for I had to tease apart which came from
``go``, which came from ``goflymake``, and which came from Emacs.

This commit introduces some extra documentation and forensic output
to assist users of this tool.